PDCL-1310 - move SRC out of tagName constants file
