+++ src/components/Personalization/dom-actions/addNonceToInlineStyleElements.js
@@ -11,7 +11,8 @@ governing permissions and limitations under the License.
 */
 
 import { selectNodes } from "../../../utils/dom";
-import { SRC, STYLE } from "../../../constants/tagName";
+import { STYLE } from "../../../constants/tagName";
+import { SRC } from "../../../constants/elementAttribute";
 import { getAttribute, getNonce } from "./dom";
 
 const is = (element, tagName) => element.tagName === tagName;
