+++ src/components/Personalization/dom-actions/images.js
@@ -11,7 +11,8 @@ governing permissions and limitations under the License.
 */
 
 import { createNode, selectNodes } from "../../../utils/dom";
-import { IMG, SRC } from "../../../constants/tagName";
+import { IMG } from "../../../constants/tagName";
+import { SRC } from "../../../constants/elementAttribute";
 import { getAttribute } from "./dom";
 
 export const isImage = element => element.tagName === IMG;
