+++ src/constants/tagName.js
@@ -16,5 +16,4 @@ export const IMG = "IMG";
 export const DIV = "DIV";
 export const STYLE = "STYLE";
 export const SCRIPT = "SCRIPT";
-export const SRC = "src";
 export const HEAD = "HEAD";
